refactor: skeleton of categories search page
