+++ src/components/ui/skeletons/index.ts
@@ -1,7 +1,7 @@
 export * from './skeleton';
 export * from './authors-search-container-skeleton';
 export * from './books-search-container-skeleton';
-export * from './categories-grid-skeleton';
+export * from './categories-search-container-skeleton';
 export * from './data-table-skeleton';
 export * from './pagination-container-skeleton';
 export * from './search-bar-skeleton';
